% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdc.R
\name{ccdc}
\alias{ccdc}
\title{Implementation of simple CCDC}
\usage{
ccdc(doy, val, fit = NULL, period = NULL)
}
\arguments{
\item{doy}{a list of julain date}

\item{val}{values corresponding to \code{doy}.}

\item{fit}{NULL or julian dates to be fit. This change return of the function.}

\item{period}{A maximum julian date to limit fitting (optional). Use NULL to ignore.}
}
\value{
If fit = NULL, coefficients of CCDC implementation. Otherwise fitted values using coefficients.
}
\description{
Implementation of a simple CCDC harmonic regression. See original article by Zhu et al. (2014) for details.
}
