% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdcTimeSeries.R
\name{ccdcTimeSeries}
\alias{ccdcTimeSeries}
\title{Implementation of CCDC for time series data}
\usage{
ccdcTimeSeries(
  values,
  VI = "NBR",
  startDOY = 2016,
  fillNA = F,
  init_rirls,
  rirls,
  ccdc,
  init_ccdc,
  rmse
)
}
\arguments{
\item{values}{a dataframe with time series Landsat data.}

\item{VI}{character. A spectral index to fit harmonic model.}

\item{startDOY}{numeric. A julian date of starting CCD change detection.}

\item{fillNA}{logical. If TRUE, fill NA values in time seires with original values.}

\item{init_rirls}{(function)}

\item{rirls}{(function)}

\item{ccdc}{(function)}

\item{init_ccdc}{(function)}

\item{rmse}{(function)}
}
\value{
a dataframe with CCDC results and potential disturbance flags.
}
\description{
Implementation of CCDC fitting and variable deriviation for time series Landsat data.
}
\details{
Main function for CCDC implementation. For Sentinel-1 data, use \code{\link{ccdcTimeSeriesSAR}}.
}
