% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdcTimeSeries.R
\name{ccdcTimeSeriesSAR}
\alias{ccdcTimeSeriesSAR}
\title{Implementation of CCDC for time series data (Sentinel-1)}
\usage{
ccdcTimeSeriesSAR(
  values,
  VI = c("VV", "VH"),
  startDOY = 2016,
  init_rirls,
  rirls,
  ccdc,
  init_ccdc,
  rmse
)
}
\arguments{
\item{values}{a dataframe with time series Landsat data.}

\item{VI}{character. A spectral index to fit harmonic model.}

\item{startDOY}{numeric. A julian day of starting CCD change detection.}

\item{init_rirls}{(function)}

\item{rirls}{(function)}

\item{ccdc}{(function)}

\item{init_ccdc}{(function)}

\item{rmse}{(function)}
}
\value{
a dataframe with CCDC results and potential disturbance flags.
}
\description{
Implementation of CCDC fitting and variable deriviation for time series Sentinel-1 data.
}
\details{
Main function for CCDC implementation. For Landsat 8 data, use \code{\link{ccdcTimeSeries}}.
}
