% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{dt_ref}
\alias{dt_ref}
\title{Sample reference data}
\format{A dataframe with 300 rows and 3 columns (x, y, date).}
\usage{
dt_ref
}
\description{
A dataframe containing reference data for dsturbance detection.
}
\keyword{datasets}
