% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractValues.R
\name{extractParallel}
\alias{extractParallel}
\title{Extract values from raster}
\usage{
extractParallel(
  ls_l8,
  l8_doys,
  dt_ref,
  col_names = paste0("B", 2:7),
  max_cores
)
}
\arguments{
\item{ls_l8}{a file list of Landsat 8 or Sentinel-1. All files should have the same extent.}

\item{l8_doys}{a list of julian day of Landsat 8 or Sentinel-1 (numeric).
This should be the same order and length as \code{ls_l8}.}

\item{dt_ref}{a dataframe of reference data for RF. This should contain column 'x', 'y', and 'date'.
x and y should indicate locations in the same crs as Landsat or Sentinel. 'date' should indicate the timing of disturbance.
If there is no disturbance at that location, use NA.}

\item{col_names}{names of bands. If L8, this should be B2, B3...B7. If S1, this should be VV and VH.}

\item{max_cores}{numeric. Maximum numbers of cores used for parallel processing.}
}
\value{
a dataframe of data values for each location.
}
\description{
Extract values from L8, S1, or DEM at locations x and y.
}
