% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractValues.R
\name{extractTopo}
\alias{extractTopo}
\title{Extract values from raster (DEM)}
\usage{
extractTopo(ls_dem, dt_ref, col_names = c("elevation", "slope"))
}
\arguments{
\item{ls_dem}{a file name of DEM with 2 bands (elevation and slope in this order).}

\item{dt_ref}{a dataframe of reference data for RF. This should contain column 'x', 'y', and 'date'.
x and y should indicate locations in the same crs as Landsat or Sentinel. 'date' should indicate the timing of disturbance.
If there is no disturbance at that location, use NA.}

\item{col_names}{names of bands. Should be elevation and slope.}
}
\value{
a dataframe of data values for each location.
}
\description{
Extract values from raster (DEM)
}
