% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractValuesCrop.R
\name{extractTopoCrop}
\alias{extractTopoCrop}
\title{Extract values from raster (DEM)}
\usage{
extractTopoCrop(ls_dem, e1, col_names = c("elevation", "slope"))
}
\arguments{
\item{ls_dem}{a file name of DEM with 2 bands (elevation and slope in this order).}

\item{e1}{an extent of cropping.}

\item{col_names}{names of bands. Should be elevation and slope.}
}
\value{
a dataframe of data values for each location.
}
\description{
Extract values from raster (DEM)
}
