% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridExtent.R
\name{gridExtent}
\alias{gridExtent}
\title{Generate grid tiles}
\usage{
gridExtent(r_use, nx = 10, ny = 10, buffer_xy = 0)
}
\arguments{
\item{r_use}{raster object to extract extent.}

\item{nx}{a number of split in x.}

\item{ny}{a number of split in y.}

\item{buffer_xy}{a buffer for overlap between tiles.}
}
\value{
a data frame (xmin, xmax, ymin, ymax) of grids split into tiles.
}
\description{
Generate grid tiles
}
