% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementRF.R
\name{implementRF}
\alias{implementRF}
\title{Build random forest model}
\usage{
implementRF(
  dt_train,
  dt_test,
  dir_rf = NULL,
  name_rdata = NULL,
  do_prallel = T
)
}
\arguments{
\item{dt_train}{dataframe. Training data for RF. Reference col should be 'manual'.}

\item{dt_test}{dataframe. Test data for RF. This should have the same columns as \code{dt_train}.}

\item{dir_rf}{a directory to save the result of RF. If NULL, a subfolder will be automatically generated.}

\item{name_rdata}{filefame for saving the RF result. If NULL, name will be automatically generated.}

\item{do_prallel}{logical. If TRUE, use parallel process to tune RF.}
}
\value{
a list of the results of RF modeling. Rdata file is saved in a subfolder.
}
\description{
Build a RF model using caret and save result in a subfolder.
}
