% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdc.R
\name{init_ccdc}
\alias{init_ccdc}
\title{initial implementation of CCDC}
\usage{
init_ccdc(doys, vals, index, col_adj, wch_start, wch_end, ccdc, rmse)
}
\arguments{
\item{doys}{a list of julain date}

\item{vals}{values corresponding to \code{doys}.}

\item{index}{cols for using ccdc.}

\item{col_adj}{adjustment for cols to insert results into dataframe.}

\item{wch_start}{numeric. Designate which is the start date of fitting in \code{doys}.}

\item{wch_end}{numeric. Designate which is the end date of fitting in \code{doys}.}

\item{ccdc}{(function)}

\item{rmse}{(function)}
}
\value{
a data frame with the result of initial CCDC implementation.
}
\description{
initial implementation of CCDC
}
