% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rirls.R
\name{init_rirls}
\alias{init_rirls}
\title{Initial implementation of RIRLS}
\usage{
init_rirls(doys, val, wch_strt, wch_end, index, rirls)
}
\arguments{
\item{doys}{a list of julain date}

\item{val}{values corresponding to \code{doy}.}

\item{wch_strt}{numeric. Designate which is the start date of fitting in \code{doys}.}

\item{wch_end}{numeric. Designate which is the end date of fitting in \code{doys}.}

\item{index}{cols for used.}

\item{rirls}{(function)}
}
\value{
a dataframe masked with NA
}
\description{
Initial implementation of RIRLS before fitting CCDC (for Landsat only)
}
\details{
Band values (surface reflectance) must be multiplied by 10000 (range from 0 to 10000).
}
