% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judgeProbSeriesDetect.R
\name{judgeProbSeriesDetect}
\alias{judgeProbSeriesDetect}
\title{Detect disturbance in time series}
\usage{
judgeProbSeriesDetect(
  dt_l8_use,
  dt_s1_use,
  ids,
  threshold,
  seq_l,
  lag_m,
  savename = NULL
)
}
\arguments{
\item{dt_l8_use}{a dataframe form Landsat 8 (or NULL). Should have 'ID', 'DOY', 'predict' columns.}

\item{dt_s1_use}{a dataframe form Sentinel-1 (or NULL). Should have 'ID', 'DOY', 'predict' columns.}

\item{ids}{vector. a list of IDs to be mapped.}

\item{threshold}{a threshold to identify disturbance.}

\item{seq_l}{a consective vector to judsge disturbance.}

\item{lag_m}{(Deprecated). lag date for identifying disturbance.}

\item{savename}{filenames for saving plots. If NULL, no plots.}
}
\value{
a dataframe with disturbance detection.
}
\description{
Detect disturbance in time series of CCDC and RF results. If \code{dt_l8_use} = NULL, then only S1 data is used for
disturbance detection.  If \code{dt_s1_use} = NULL, then only L8 data is used for disturbance detection.
}
