% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{ls_dem}
\alias{ls_dem}
\title{Sample DEM data}
\format{A list of DEM (elevation, slope) RasterStack.}
\usage{
ls_dem
}
\description{
A list containing \code{RasterStack} of DEM (2 bands).
}
\keyword{datasets}
