% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{ls_l8}
\alias{ls_l8}
\title{Sample Landsat 8 data}
\format{A list of 75 Landsat 8 (SR data) RasterStack.}
\usage{
ls_l8
}
\description{
A list containing \code{RasterStack} of Landsat 8 (6 bands).
}
\keyword{datasets}
