% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{ls_s1}
\alias{ls_s1}
\title{Sample Sentinel-1 data}
\format{A list of 98 Sentinel-1 (VV, VH) RasterStack.}
\usage{
ls_s1
}
\description{
A list containing \code{RasterStack} of Sentinel-1 (2 bands).
}
\keyword{datasets}
