% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDisturbanceL8S1.R
\name{mapDisturbanceL8S1}
\alias{mapDisturbanceL8S1}
\title{Map forest disturbance using Landsat 8 and Sentinel-1}
\usage{
mapDisturbanceL8S1(
  ls_l8,
  ls_s1,
  l8_doys,
  s1_doys,
  dt_ref,
  ls_dem = NULL,
  dir_save,
  VI,
  rf_model = NULL,
  startDOY,
  endDOY,
  mmu = NULL,
  only_rf = F,
  max_cores = 20,
  threshold = 0.5
)
}
\arguments{
\item{ls_l8}{a file list of Landsat 8 with 6 bands. All files should have the same extent.}

\item{ls_s1}{a file list of Sentinel-1 with 2 bands (VV and VH, in this order). If this is used with Landsat,
all files should have the same extent as Landsat (i.e. same origin, crs and resolution)}

\item{l8_doys}{a list of julian day of Landsat 8 (numeric). This should be the same order and length as \code{ls_l8}.}

\item{s1_doys}{a list of julian day of Sentinel-1 (numeric). This should be the same order and length as \code{ls_s1}.}

\item{dt_ref}{a dataframe of reference data for RF. This should contain column 'x', 'y', and 'date'. x and y should
indicate locations in the same crs as Landsat or Sentinel. 'date' should indicate the timing of disturbance.
If there is no disturbance at that location, use NA.}

\item{ls_dem}{a file name of DEM with 2 bands (elevation and slope in this order). This is used for RF model.
If this is not to be used for model, set NULL.}

\item{dir_save}{a directory for save the results.}

\item{VI}{names of spectral index used for Landsat.}

\item{rf_model}{a list of two RF models (as list) to skip RF model. Set NULL to build RF model.}

\item{startDOY}{a julian day of the initial timing of disturbance detection (e.g. 2016-02-01 should be 2016.08767)}

\item{endDOY}{a julian day of the terminate day of disturbance detection.}

\item{mmu}{numeric. minimum mapping unit (pixel) for final maps. New tif image will be additionally generated.}

\item{only_rf}{logical. If TRUE, only build RF models. If FALSE, build RF models and map disturbance detection.}

\item{max_cores}{numeric. Maximum numbers of cores used for parallel processing.}

\item{threshold}{numeric from 0 to 1. A threshold to detect disturbance from time series disturbance probabilities.}
}
\value{
a filename of mapped disturbance.
}
\description{
Detect and map forest disturbance using Landsat 8 and Sentinel-1 time series.
}
\details{
This function can be mainly used for three ways.
1. build RF models and then detect & map disturbance.
2. build RF models (no mapping nor detection of disturbance)
3. detect & map disturbance (mandatory with RF models)

The use of both Landsat 8 and Sentinel-1 is full implementation of this algorithm.
However, you can use either Landsat 8 or Sentinle-1 if you want. The process takes very long time.
This use parallel process to save time.
}
\note{
Only 50\% of sample ID is used for RF modeling.Subfolders are automatically genereted for saving RF models
and mapping temporally tiles of a disturbance map.
}
\examples{

}
