% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPatchMMU.R
\name{mapPatchMMU}
\alias{mapPatchMMU}
\title{apply MMU for disturbance map}
\usage{
mapPatchMMU(dir_map, mmu = 3, band = 1)
}
\arguments{
\item{dir_map}{a filename of disturbance map.}

\item{mmu}{a minimum mapping unit in pixels.}

\item{band}{a band for applying mmu.}
}
\value{
a disturbance map with mmu. A new disturbance map is also generated.
}
\description{
Remove a small patches using a minimum mapping unit.
}
