% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judgeProbSeriesDetect.R
\name{matchVector}
\alias{matchVector}
\title{detect consective anomalies---}
\usage{
matchVector(x, pattern)
}
\arguments{
\item{x}{a vector of logical.}

\item{pattern}{a pattern of matching vector. Should be a vector of ligical.}
}
\description{
detect consective anomalies---
}
