% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementRF.R
\name{sinkRFresult}
\alias{sinkRFresult}
\title{sink the result of RF}
\usage{
sinkRFresult(rf_tune, mat_conf, filename)
}
\arguments{
\item{rf_tune}{tune result of RF.}

\item{mat_conf}{confusion matrix.}

\item{filename}{a filename to save.}
}
\description{
sink the result of RF
}
